Extract validation helper fns
